#55 Move runtime functions to typing module
