+++ xarray_dataclasses/parser.py
@@ -3,18 +3,24 @@ __all__ = ["parse"]
 
 # standard library
 from dataclasses import dataclass, Field
-from itertools import chain
-from typing import Any, Dict, ForwardRef, List, Optional, Tuple, TypeVar
+from typing import Any, Dict, List, Optional, Tuple, TypeVar
 
 
 # third-party packages
 import numpy as np
 import xarray as xr
-from typing_extensions import Annotated, get_args, get_origin, Literal, Protocol
+from typing_extensions import get_args, Protocol
 
 
 # submodules
-from .typing import ArrayLike, DataClassLike, FieldType
+from .typing import (
+    ArrayLike,
+    DataClassLike,
+    FieldType,
+    get_dims,
+    get_dtype,
+    unannotate,
+)
 
 
 # type hints
@@ -59,8 +65,8 @@ class DataArray:
     @classmethod
     def from_field(cls, field: Field[Any], value: Any) -> "DataArray":
         """Create an instance from a Coord/Data-type field and a value."""
-        dims, dtype = get_args(get_args(unannotate(field.type))[0])
-        type = dict(dims=parse_dims(dims), dtype=parse_dtype(dtype))
+        t_dims, t_dtype = get_args(get_args(unannotate(field.type))[0])
+        type = dict(dims=get_dims(t_dims), dtype=get_dtype(t_dtype))
         return cls(field.name, type, value)
 
     def instantiate(self) -> xr.DataArray:
@@ -141,56 +147,6 @@ def parse(dataclass: DataClassLike) -> ParsedDataClass:
 
 
 # helper features
-def parse_dims(type_like: Any) -> Dims:
-    """Parse a type-like object and get dims."""
-    type_like = unannotate(type_like)
-
-    if type_like == () or type_like is NoneType:
-        return ()
-
-    if isinstance(type_like, ForwardRef):
-        return (type_like.__forward_arg__,)
-
-    if isinstance(type_like, str):
-        return (type_like,)
-
-    origin = get_origin(type_like)
-    args = get_args(type_like)
-
-    if origin is tuple:
-        return tuple(chain(*map(parse_dims, args)))
-
-    if origin is Literal:
-        return tuple(map(str, args))
-
-    raise ValueError(f"Could not parse {type_like}.")
-
-
-def parse_dtype(type_like: Any) -> Dtype:
-    """Parse a type-like object and get dtype."""
-    type_like = unannotate(type_like)
-
-    if type_like is Any or type_like is NoneType:
-        return None
-
-    if isinstance(type_like, type):
-        return type_like.__name__
-
-    if isinstance(type_like, ForwardRef):
-        return type_like.__forward_arg__
-
-    if isinstance(type_like, str):
-        return type_like
-
-    origin = get_origin(type_like)
-    args = get_args(type_like)
-
-    if origin is Literal and len(args) == 1:
-        return str(args[0])
-
-    raise ValueError(f"Could not parse {type_like}.")
-
-
 def to_dataarray(
     data: Any,
     dims: Dims,
@@ -225,25 +181,3 @@ def to_subspace(dataarray: xr.DataArray, dims: Dims) -> xr.DataArray:
     """Return the subspace of a DataArray with given dims."""
     indexers = {dim: 0 for dim in dataarray.dims if dim not in dims}
     return dataarray.isel(indexers)
-
-
-def unannotate(obj: T) -> T:
-    """Recursively remove Annotated types."""
-    if get_origin(obj) is Annotated:
-        obj = get_args(obj)[0]
-
-    origin = get_origin(obj)
-
-    if origin is None:
-        return obj
-
-    args = map(unannotate, get_args(obj))
-    args = tuple(filter(None, args))
-
-    try:
-        return origin[args]
-    except TypeError:
-        import typing
-
-        name = origin.__name__.capitalize()
-        return getattr(typing, name)[args]
